Refactored code into parseutil/misc.py
